+++ jsonrpc/endpoints_debug.go
@@ -2,21 +2,17 @@ package jsonrpc
 
 import (
 	"context"
-	"encoding/hex"
 	"encoding/json"
 	"errors"
 	"fmt"
 	"net/http"
 	"sort"
-	"strings"
 	"sync"
 	"time"
 
 	"github.com/0xPolygonHermez/zkevm-node/jsonrpc/types"
 	"github.com/0xPolygonHermez/zkevm-node/log"
 	"github.com/0xPolygonHermez/zkevm-node/state"
-	"github.com/0xPolygonHermez/zkevm-node/state/runtime/fakevm"
-	"github.com/0xPolygonHermez/zkevm-node/state/runtime/instrumentation"
 	"github.com/ethereum/go-ethereum/common"
 	ethTypes "github.com/ethereum/go-ethereum/core/types"
 	"github.com/jackc/pgx/v4"
@@ -54,27 +50,6 @@ type traceConfig struct {
 	EnableReturnData bool            `json:"enableReturnData"`
 	Tracer           *string         `json:"tracer"`
 	TracerConfig     json.RawMessage `json:"tracerConfig"`
-}
-
-// StructLogRes represents the debug trace information for each opcode
-type StructLogRes struct {
-	Pc            uint64             `json:"pc"`
-	Op            string             `json:"op"`
-	Gas           uint64             `json:"gas"`
-	GasCost       uint64             `json:"gasCost"`
-	Depth         int                `json:"depth"`
-	Error         string             `json:"error,omitempty"`
-	Stack         *[]types.ArgBig    `json:"stack,omitempty"`
-	Memory        *[]string          `json:"memory,omitempty"`
-	Storage       *map[string]string `json:"storage,omitempty"`
-	RefundCounter uint64             `json:"refund,omitempty"`
-}
-
-type traceTransactionResponse struct {
-	Gas         uint64         `json:"gas"`
-	Failed      bool           `json:"failed"`
-	ReturnValue interface{}    `json:"returnValue"`
-	StructLogs  []StructLogRes `json:"structLogs"`
 }
 
 type traceBlockTransactionResponse struct {
@@ -301,11 +276,6 @@ func (d *DebugEndpoints) buildTraceTransaction(ctx context.Context, hash common.
 		traceCfg = defaultTraceConfig
 	}
 
-	// check tracer
-	if traceCfg.Tracer != nil && *traceCfg.Tracer != "" && !isBuiltInTracer(*traceCfg.Tracer) && !isJSCustomTracer(*traceCfg.Tracer) {
-		return RPCErrorResponse(types.DefaultErrorCode, "invalid tracer", nil, false)
-	}
-
 	stateTraceConfig := state.TraceConfig{
 		DisableStack:     traceCfg.DisableStack,
 		DisableStorage:   traceCfg.DisableStorage,
@@ -322,134 +292,7 @@ func (d *DebugEndpoints) buildTraceTransaction(ctx context.Context, hash common.
 		return nil, types.NewRPCError(types.DefaultErrorCode, errorMessage)
 	}
 
-	// if a tracer was specified, then return the trace result
-	if stateTraceConfig.Tracer != nil && *stateTraceConfig.Tracer != "" && len(result.ExecutorTraceResult) > 0 {
-		return result.ExecutorTraceResult, nil
-	}
-
-	receipt, err := d.state.GetTransactionReceipt(ctx, hash, dbTx)
-	if err != nil {
-		const errorMessage = "failed to tx receipt"
-		log.Errorf("%v: %v", errorMessage, err)
-		return nil, types.NewRPCError(types.DefaultErrorCode, errorMessage)
-	}
-
-	failed := receipt.Status == ethTypes.ReceiptStatusFailed
-	var returnValue interface{}
-	if stateTraceConfig.EnableReturnData {
-		returnValue = common.Bytes2Hex(result.ReturnValue)
-	}
-
-	structLogs := d.buildStructLogs(result.StructLogs, *traceCfg)
-
-	resp := traceTransactionResponse{
-		Gas:         result.GasUsed,
-		Failed:      failed,
-		ReturnValue: returnValue,
-		StructLogs:  structLogs,
-	}
-
-	return resp, nil
-}
-
-func (d *DebugEndpoints) buildStructLogs(stateStructLogs []instrumentation.StructLog, cfg traceConfig) []StructLogRes {
-	structLogs := make([]StructLogRes, 0, len(stateStructLogs))
-	memory := fakevm.NewMemory()
-	for _, structLog := range stateStructLogs {
-		errRes := ""
-		if structLog.Err != nil {
-			errRes = structLog.Err.Error()
-		}
-
-		op := structLog.Op
-		if op == "SHA3" {
-			op = "KECCAK256"
-		} else if op == "STOP" && structLog.Pc == 0 {
-			// this stop is generated for calls with single
-			// step(no depth increase) and must be ignored
-			continue
-		}
-
-		structLogRes := StructLogRes{
-			Pc:            structLog.Pc,
-			Op:            op,
-			Gas:           structLog.Gas,
-			GasCost:       structLog.GasCost,
-			Depth:         structLog.Depth,
-			Error:         errRes,
-			RefundCounter: structLog.RefundCounter,
-		}
-
-		if !cfg.DisableStack {
-			stack := make([]types.ArgBig, 0, len(structLog.Stack))
-			for _, stackItem := range structLog.Stack {
-				if stackItem != nil {
-					stack = append(stack, types.ArgBig(*stackItem))
-				}
-			}
-			structLogRes.Stack = &stack
-		}
-
-		if cfg.EnableMemory {
-			memory.Resize(uint64(structLog.MemorySize))
-			if len(structLog.Memory) > 0 {
-				memory.Set(uint64(structLog.MemoryOffset), uint64(len(structLog.Memory)), structLog.Memory)
-			}
-
-			if structLog.MemorySize > 0 {
-				// Populate the structLog memory
-				structLog.Memory = memory.Data()
-
-				// Convert memory to string array
-				const memoryChunkSize = 32
-				memoryArray := make([]string, 0, len(structLog.Memory))
-
-				for i := 0; i < len(structLog.Memory); i = i + memoryChunkSize {
-					slice32Bytes := make([]byte, memoryChunkSize)
-					copy(slice32Bytes, structLog.Memory[i:i+memoryChunkSize])
-					memoryStringItem := hex.EncodeToString(slice32Bytes)
-					memoryArray = append(memoryArray, memoryStringItem)
-				}
-
-				structLogRes.Memory = &memoryArray
-			} else {
-				memory = fakevm.NewMemory()
-				structLogRes.Memory = &[]string{}
-			}
-		}
-
-		if !cfg.DisableStorage && len(structLog.Storage) > 0 {
-			storage := make(map[string]string, len(structLog.Storage))
-			for storageKey, storageValue := range structLog.Storage {
-				k := hex.EncodeToString(storageKey.Bytes())
-				v := hex.EncodeToString(storageValue.Bytes())
-				storage[k] = v
-			}
-			structLogRes.Storage = &storage
-		}
-
-		structLogs = append(structLogs, structLogRes)
-	}
-	return structLogs
-}
-
-// isBuiltInTracer checks if the tracer is one of the
-// built-in tracers
-func isBuiltInTracer(tracer string) bool {
-	// built-in tracers
-	switch tracer {
-	case "callTracer", "4byteTracer", "prestateTracer", "noopTracer":
-		return true
-	default:
-		return false
-	}
-}
-
-// isJSCustomTracer checks if the tracer contains the
-// functions result and fault which are required for a custom tracer
-// https://geth.ethereum.org/docs/developers/evm-tracing/custom-tracer
-func isJSCustomTracer(tracer string) bool {
-	return strings.Contains(tracer, "result") && strings.Contains(tracer, "fault")
+	return result.TraceResult, nil
 }
 
 // waitTimeout waits for the waitGroup for the specified max timeout.
