+++ state/runtime/runtime.go
@@ -242,15 +242,14 @@ var (
 // ExecutionResult includes all output after executing given evm
 // message no matter the execution itself is successful or not.
 type ExecutionResult struct {
-	ReturnValue         []byte // Returned data from the runtime (function result or data supplied with revert opcode)
-	GasLeft             uint64 // Total gas left as result of execution
-	GasUsed             uint64 // Total gas used as result of execution
-	Err                 error  // Any error encountered during the execution, listed below
-	CreateAddress       common.Address
-	StateRoot           []byte
-	StructLogs          []instrumentation.StructLog
-	ExecutorTrace       instrumentation.ExecutorTrace
-	ExecutorTraceResult json.RawMessage
+	ReturnValue   []byte // Returned data from the runtime (function result or data supplied with revert opcode)
+	GasLeft       uint64 // Total gas left as result of execution
+	GasUsed       uint64 // Total gas used as result of execution
+	Err           error  // Any error encountered during the execution, listed below
+	CreateAddress common.Address
+	StateRoot     []byte
+	FullTrace     instrumentation.FullTrace
+	TraceResult   json.RawMessage
 }
 
 // Succeeded indicates the execution was successful
