+++ state/types.go
@@ -71,10 +71,8 @@ type ProcessTransactionResponse struct {
 	ChangesStateRoot bool
 	// Tx is the whole transaction object
 	Tx types.Transaction
-	// ExecutionTrace contains the traces produced in the execution
-	ExecutionTrace []instrumentation.StructLog
-	// CallTrace contains the call trace.
-	CallTrace instrumentation.ExecutorTrace
+	// FullTrace contains the call trace.
+	FullTrace instrumentation.FullTrace
 	// EffectiveGasPrice effective gas price used for the tx
 	EffectiveGasPrice string
 	//EffectivePercentage effective percentage used for the tx
